Desktop/rent-label-api-client/requirements.txt
fastapi
uvicorn
httpx
Pillow
pytesseract
python-multipart
